BaseStyles: + simple darkmod button functionality
