+++ source/css/scaffolding.css
@@ -2,6 +2,8 @@
 *::before,
 *::after {
   box-sizing: border-box;
+
+  transition: color, background-color 500ms;
 }
 
 :focus {
