+++ source/js/main.js
@@ -1,0 +1,3 @@
+import init from './app'
+
+init()
