+++ source/js/theme-button/common.js
@@ -1,0 +1,6 @@
+const ColorThemeTypes = {
+  DARK: `dark`,
+  LIGHT: `light`,
+}
+
+export { ColorThemeTypes }
