+++ source/js/theme-button/helpers.js
@@ -1,0 +1,5 @@
+const setColorTheme = (color) => {
+  document.documentElement.setAttribute(`data-theme`, color)
+}
+
+export { setColorTheme }
